+++ src/YarleOptions.ts
@@ -1,7 +1,8 @@
-import { CharacterMap } from 'CharacterMap';
-import { TagSeparatorReplaceOptions } from './models';
 import { OutputFormat } from './output-format';
+import { TagSeparatorReplaceOptions, SearchAndReplace } from './models';
 import { TaskOutputFormat } from './task-output-format';
+import { CharacterMap } from 'CharacterMap';
+import { ImageSizeFormat } from 'image-size-format';
 
 export interface YarleOptions {
     enexDir?: string; // used by command line
@@ -48,7 +49,8 @@ export interface YarleOptions {
     monospaceIsCodeBlock?: boolean;
     dateFormat?: string;
     nestedTags?: TagSeparatorReplaceOptions;
-    keepImageSize?: OutputFormat;
+    imageSizeFormat?: ImageSizeFormat;
+    keepImageSize?: boolean;
     keepOriginalAmountOfNewlines?: boolean;
     generateNakedUrls?: boolean;
     addExtensionToInternalLinks?: boolean;
@@ -59,5 +61,7 @@ export interface YarleOptions {
     obsidianTaskTag?: string;
     useUniqueUnknownFileNames?: boolean;
     useLevenshteinForLinks?: boolean;
-    keepFontColors?: boolean;
+    keepEvernoteLinkIfNoNoteFound?: boolean;
+    convertColorsToMDHighlight?: boolean;
+    globalReplacementSettings?: Array<SearchAndReplace>;
 }