+++ src/image-size-format.ts
@@ -1,0 +1,4 @@
+export const enum ImageSizeFormat {
+    ObsidianMD= 'ObsidianMD',
+    StandardMD= 'StandardMD'
+}