+++ src/utils/save-tana-file.ts
@@ -3,8 +3,9 @@ import { RuntimePropertiesSingleton } from './../runtime-properties';
 import { writeFile } from './file-utils';
 import { getJsonFilePath } from './folder-utils';
 import { loggerInfo } from './loggerInfo';
+import { EvernoteNoteData } from "./../models";
 
-export const saveTanaFile = (tanaNote: TanaIntermediateFile, note: any) => {
+export const saveTanaFile = (tanaNote: TanaIntermediateFile, note: EvernoteNoteData) => {
 
     const absJsonFilePath = getJsonFilePath(note);
     const runtimeProps = RuntimePropertiesSingleton.getInstance();