+++ src/utils/tana/convert-to-tana-node.ts
@@ -107,7 +107,8 @@ export const convertChild = (data: NoteData, child: string) => {
 }
 
 export const convert2TanaNode = (data: NoteData):TanaIntermediateFile => {
-    const firstLevelChildren = data.content.replace(/(\n|\r\n)+/g, '\n').split('\n')
+    const firstLevelChildren = data.appliedMarkdownContent.replace(/(\n|\r\n)+/g, '\n').split('\n')
+    
     const rootContent = data.title
     
     const childrenNodes: Array<TanaIntermediateNode> = [];